pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "bharat-market-pulse"
version = "0.1.0"
description = "Indian market portfolio intelligence pipeline"
authors = [{ name = "Project Nova" }]
dependencies = [
  "pydantic",
  "python-dotenv",
  "requests",
  "pandas",
  "numpy",
  "pillow",
  "easyocr",
  "pytesseract",
]

[project.optional-dependencies]
dev = [
  "pytest",
  "ruff",
  "black",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"
